dev/ttrss_notion_sync_app/requirements.txt
psycopg2-binary==2.9.6
notion-client==2.0.0
tqdm==4.65.0
python-dotenv==1.0.0
